Refactor codebase for workflow packages
