+++ src/myteam/workflow/models.py
@@ -1,0 +1,65 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Any, TypedDict
+
+
+class AgentConfig(TypedDict):
+    name: str
+    argv: list[str]
+    exit_text: str | None
+
+
+class StepDefinition(TypedDict, total=False):
+    prompt: str
+    output: dict[str, Any]
+    input: Any
+    agent: str
+
+
+class WorkflowStepEntry(TypedDict):
+    name: str
+    definition: StepDefinition
+
+
+class WorkflowDefinition(TypedDict):
+    steps: list[WorkflowStepEntry]
+
+
+class CompletedStepState(TypedDict, total=False):
+    prompt: str
+    input: Any
+    agent: str | None
+    output: Any
+
+
+WorkflowOutput = dict[str, CompletedStepState]
+
+
+@dataclass
+class RunContext:
+    prior_steps: dict[str, Any]
+    default_agent: str
+
+
+@dataclass
+class StepResult:
+    step_name: str
+    status: str
+    output: Any | None = None
+    error_type: str | None = None
+    error_message: str | None = None
+    transcript: str = ""
+
+
+@dataclass
+class WorkflowRunResult:
+    status: str
+    output: WorkflowOutput | None = None
+    failed_step_name: str | None = None
+
+
+@dataclass
+class PtyRunResult:
+    exit_code: int
+    transcript: str = ""
